Optimize benchmark code structure
